print profiles: manually set extrusion width to match nozzle
slic3r does a strange computation of how much filament to extrude vs.
line spacing.

The extrusion width specifies how wide you want the squished extrusion
to be.  That is the width that will be filled by a rectangle
layer_height high with 2 half circles on each end that are radius
layer_height / 2.

Based on the extrusion width it then makes the line spacing

nozzle_diameter + layer_height * (1 - pi/4)

Each of the default profiles now sets its extrusion width so that
the line spacing is now nozzle_diameter.