idex: drop accel and increase x motor current
I had layer shifts when printing a benchy.

It was stupid to increase acceleration when I was still trying to get
this printer working, go back to something that was safe and working
on other printers.

I only dropped the motor current because I was trying to get stall
detection working.  Put it back where it was.