print: default: try to improve dual extruder support print settings
Switch back to grid because organic supports don't actually come close
to the print right now!!  I am seeing 0.6mm offset between the support
and the print it is supposed to be supporting.

Increase the interface layers to try to create a smoother surface to
support the first printed layer on.

Decrease the xy gap of the support to try to remove sagging.