ender-5:  Home Z after bed levelling and klicky probe retraction
If I decide to use the mesh bed compensation, I need the Z height to
be set relative to the same position that the compensation was measured
at.

In the same block of gcode I also added the forced retract for the
klicky probe to unsure it is never loaded when we try to print.