v-king: Tweak filament sensor config
I was getting random filament errors and some crazy readings (like
a max of 550%).  Make the measured range larger (9mm -> 22mm) to
make a small error in the readings not make the value go crazy.

Also make the range larger since we are really trying to catch
filament running out as the only case that we're likely to save the
print anyway.  Tangled filament, nozzle clog, etc. are likely to
leave a gap in the print anyway.