rrf: taz6: klicky-pre-dock doesn't need to be so careful for z height
Previously it made sense to do the second pass on the z home button
because we were actually setting the Z zero.  But, now that we're
just using it to go to the probe height without changing the Z zero,
we can just do the first pass touch off the button and call it close
enough.