rrf: prepare-to-print: use T0 to set Z for IDEX prints
When printing with both nozzles, we calibrate the Z offset of
T1 relative to T0.  Previously we were then probing the bed using
T1 which sets the Z=0 value to be the machine position when
T1 hits the bed.

That is, if T1 is 1mm higher than T0 we probe and set Z=0 to the
point at which T1 hits the bed.  T0 will now try to print 1mm too
low because it is 1mm lower than T1.

Fix this by making sure we do G32/G28 Z using T0 when printing
with both hotends.