rrf: idex: Allow z to travel down to -5
As far as I can tell, RRF 3.6.0 changed the behaviour so that
the Z is capped using the machinePosition in addition to the
userPosition (or some silliness).

I found when printing that if I had a tool offset of 0.35mm,
a Z lower limit of 0 then printing with T1 would print at
0.35mm (with slight changes based on the mesh compensation?).

Either way, this seems to fix the problem I was having of the
first layer of T1 being wacked out.