slic3r: Toolchange improvements
Use the prusaslicer-toolchange.g to make the following improvements
to the toolchange.

When changing tools we no longer return to the position prior to
the tool change.  Instead, PrusaSlicer already generates the gcode
to move to the next location to print.  This improves:

  * No longer drag across the print when moving to the next print
    location.  Previously we return the Z to printing height
    before the move to the next location.  Now we keep the height
    raised and it moves like a normal zhop move.

  * Less oozing of filament as we avoid a potentially large move
    across the bed by going directly to the next location.

Change the retract/extra restart on toolchange options.  Just
retract 1mm when free'ing the tool but do 2mm when returning
from the wipe position to the next printing position.  This
potentially avoids clogging the nozzle by retracting too far but
keeps the same 2mm retract when moving to the next print position.
Hopefully this causes less blobbing on the print.